Support lazily decoding pointers
Instead of immediately decoding the value pointed to, this allows lazy decoding by defining a getter. Obviously it only works when the pointer is contained within a Struct, not an Array or something else. When used, it can speed up decoding quite a bit when not all data is needed right away.